few changes
