+++ atenxa/__init__.py
@@ -1,22 +1,20 @@
 # -*- coding: utf-8 -*-
 """ATENXA - Advanced Tools for Enhancing VRMNX by Akagi
 
-VRMNXpyの，シンプル・使いやすい・リアルな拡張です。
+VRMNXの機能を拡張する，シンプル・使いやすい・リアルなツールキットです。
+
 使いやすいイベントシステムや閉塞制御などを提供します。
+大半の機能は，1行の記述で組み込めるように設計されています。
 """
 
-__version__ = '0.1.0-alpha'
+__version__ = '1.0.0-RC2'
 __date__ =    '2020/9/20'
 __author__ =  'AKAGI Rails'
 
 _DEBUG = True # デバッグモード
 
-from .basis import *
+from ._basis import *
 #from .richevent import *
 #from sprite import *
 
 #__all__ = ['basis', 'richevent']
-
-def sensor_approach(obj,ev,param):
-    printLOG('Tmp User Handler.')
-
